% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_regression.R
\name{mse_order}
\alias{mse_order}
\title{mse_order}
\usage{
mse_order(x, y)
}
\arguments{
\item{x}{a design matrix}

\item{y}{a response vector}
}
\value{
an ordered matrix x
}
\description{
given a design matrix x and response vector y, puts x's columns in order from least to most MSE.  Puts intercept 1st. Centers variables first.
}
