% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_regression.R
\name{regress_step_helper}
\alias{regress_step_helper}
\title{regress_step_helper}
\usage{
regress_step_helper(vec)
}
\arguments{
\item{vec}{response vector being to be regressed out}
}
\description{
performs some matrix algebra on a vector as preparation to regress a variable out of x or y
}
\note{
full function would be vec X solve(t(vec) X vec) X t(vec) X (vec or y); where X subs for matrix multiplication
}
