% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_plotter.R
\name{res_plotter}
\alias{res_plotter}
\title{res_plotter}
\usage{
res_plotter(res, y, xlab = "y-values", ylab = "residuals", ...)
}
\arguments{
\item{res}{a matrix of residuals}

\item{y}{the response vector}

\item{xlab}{label for x-axis, defaults to "y-values"}

\item{ylab}{label for y-axis, defaults to "residuals"}

\item{...}{additional graphical parameters for use in title() function, such as main="Title of Plot"}
}
\description{
creates a y-values vs residuals plot, with reach column of res filling one sub-plot
}
