% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_regression.R
\name{step_regression}
\alias{step_regression}
\title{step_regression}
\usage{
step_regression(x, y = NULL, labels = NULL, addIntercept = TRUE)
}
\arguments{
\item{x}{a matrix or variable that can be turned into a matrix via as.matrix.}

\item{y}{the response vector.  Defaults to NULL and redefined as the first column of x, or if that is constant, the last column of x}

\item{labels}{a list of column labels.  Defaults to NULL and redefind as the existing column names}

\item{addIntercept}{adds a column of ones as the first row of x to serve as an intercept.  Defaults to TRUE.  The 1s are added before missing ys are handled.}
}
\description{
this function takes a design matrix x, and a response vector y
}
